Resource Manager Assignment + runs with no errors!!!!!

+++ Source/Engine/Core/Time.cpp
@@ -10,7 +10,7 @@ namespace ringo {
 		m_time = duration.count() / static_cast<float>(clock_duration::period::den);
 
 		duration = clock::now() - m_frameTime;
-		m_deltaTime = duration.count();
+		m_deltaTime = (float)duration.count();
 		m_deltaTime = Min(m_deltaTime, 0.25f);
 
 		m_frameTime = clock::now();
